+++ app/src/main/java/com/tustar/demo/data/DemoRepository.kt
@@ -2,8 +2,8 @@ package com.tustar.demo.data
 
 class DemoRepository private constructor(private val demoDao: DemoDao) {
 
-    suspend fun getDemosByGroupId(groupId: Long) = demoDao.getDemosByGroupId(groupId)
-    suspend fun getDemosByParentId(parentId: Long) = demoDao.getDemosByParentId(parentId)
+    suspend fun getDemosByGroupId(groupId: Int) = demoDao.getDemosByGroupId(groupId)
+    suspend fun getDemosByParentId(parentId: Int) = demoDao.getDemosByParentId(parentId)
 
     companion object {
         @Volatile
